requirements.txt
pystray
Pillow
keyboard
pyperclip
pyinstaller
